npc: Add to structure of characters and spawn npc
The current code is quite messy and has to be optimized. Also this is still not functioning as is intended for this game.
There is also a lot of duplicated code that has to be removed!
This will still make it to a release since there were quite some changes by now and one can at least play soccer with a slime :)