physics: Improve colliders and fix rotation
Note: The crate choices I have made so far seem very correct.

I have tried the following and figured out that they are not working well for my use case(and have spent a lot of hours with them :/):
- avian
- leafwing-input-manager
- bevy-tnua

This does not mean that they are bad, this mainly serves as a note for me. tnua does not support top-down 2d which I sadly didn't figure out until I wasted a lot of time with it. For other use cases however, these might be superior to my crate choices so far.