building: Disable dynamic linking
It has caused multiple linker failures in the past, which I could only resolve with cargo clean or sometimes they resolved without any real reason. This behavior is undesirable and I don't even think that the time this feature saves is worth it compared to the time lost because of that. This might very well also be a configuration issue, but I didn't find any info on what in my setup might be at fault here.