Fix: Add Z-level offset to player for shadow
This also fixes various somewhat related oversights.

This solution is not ideal. This generally behaves kind of weird in my opinion. However, the current solution works and I don't think a similar workaround is required for npcs. I'd have added one, but the current solution just adds an offset to the player so adding an offset to the npcs would make that ineffective.

Observations:
- Adding a YSort to just visual_bundle and shadow_bundle, not container_bundle will result in the player being at a higher Z-level than the npcs. I'm pretty sure that that behavior isn't consistent since it also was the exact opposite at least once. Don't ask me why.
- Adding a YSort to just the container_bundle and the shadow_bundle will result in the shadow being at a higher Z-level than the sprite for Player and Npc.

Sadly I couldn't actually conclude anything meaningful from my testing but the behavior in general seems pretty buggy. My changes however consistently work well, it's just that all bundles need YSort and just as a sidenote, higher Z-levels than I am using right now in level module or changing the Z-level range of the camera is also definitely not working as someone would expect it to work.

This is just here to later serve me as a reference if I ever need it.